Split the pad evdev nodes out of the generic reader
evdev.rs decoded three unrelated device classes in one file. The DualSense
touchpad and motion nodes move to evdev/pad.rs behind a `Pad` enum: probe,
read, flush, release, kind. Device carries one `Option<Pad>` instead of two
Options documented as never both set, so the invariant is the type's.

Also reunites is_tv_builtin with the doc comment that had drifted onto the
predicate below it.
